Remove "Skip setting unsupported partition flag" message
This message would be displayed as part of the details for an operation
to create or format a partition as LVM2 PV on a disk using partition
table types dvh or pc98, which don't have an lvm flag.

 v Format /dev/sdb1 as lvm2 pv
   > calibrate /dev/sdb1
   > clear old file system signatures in /dev/sdb1
   > set partition type on /dev/sdb1
       Skip setting unsupported partition flag: lvm
   > create new lvm2 pv file system

Translators found this message difficult to translate, as reported in
bug 752901 - Add translator comment or fix string.  It is also only
standard practice to set the lvm flag on partitions containing LVM2 PVs,
and not required.  In hindsight reporting that an operation was skipped,
when the operation is not a necessity, is probably bad practice as it
could lead to uncertainty on behalf of the user and questions as to why
the flag isn't supported.  Just remove the message.