Ensure DMRaid /dev entries are created before using named devices
For probed DMRaid devices (when not using libparted DMRaid support)
GParted waits up to 1 second for udev to have processed all events and
created the /dev entries after starting each DMRaid array.  This was
added by this commit from 2009-09-02:
    e7352a5000ddfc7666861e3e8ef56b88510b449d
    Ensure /dev file system device entries created before adding device

Do the same for devices named on the command line too.